add sbutensor_client
