+++ neurons/validators/src/clients/compute_client.py
@@ -51,8 +51,8 @@ from pydantic import BaseModel
 from websockets.asyncio.client import ClientConnection
 
 from core.config import settings
-from core.validator import Validator
 from core.utils import _m, get_extra_info
+from clients.subtensor_client import SubtensorClient
 from services.const import GPU_MODEL_RATES
 from services.miner_service import MinerService
 from services.redis_service import (
@@ -89,7 +89,7 @@ class ComputeClient:
         self.miner_driver_awaiter_task = asyncio.create_task(self.miner_driver_awaiter())
         # self.heartbeat_task = asyncio.create_task(self.heartbeat())
         self.miner_service = miner_service
-        self.validator = Validator()
+        self.subtensor_client = SubtensorClient.get_instance()
         self.message_queue = []
         self.lock = asyncio.Lock()
 
@@ -142,7 +142,6 @@ class ComputeClient:
         return self.keypair.ss58_address
 
     async def run_forever(self) -> NoReturn:
-        asyncio.create_task(self.validator.warm_up_subtensor())
         asyncio.create_task(self.handle_send_messages())
         asyncio.create_task(self.subscribe_mesages_from_redis())
         asyncio.create_task(self.poll_rented_machines())
@@ -558,7 +557,7 @@ class ComputeClient:
             return
 
     async def get_miner_axon_info(self, hotkey: str) -> bittensor.AxonInfo:
-        miners = self.validator.fetch_miners()
+        miners = self.subtensor_client.fetch_miners()
         neurons = [n for n in miners if n.hotkey == hotkey]
         if not neurons:
             raise ValueError(f"Miner with {hotkey=} not present in this subnetwork")
