+++ neurons/validators/src/validator.py
@@ -13,10 +13,7 @@ wait_for_services_sync()
 
 
 async def app_lifespan(app: FastAPI):
-    if settings.DEBUG:
-        validator = Validator(debug_miner=settings.get_debug_miner())
-    else:
-        validator = Validator()
+    validator = Validator()
     # Run the miner in the background
     task = asyncio.create_task(validator.start())
